Get all math-input tests passing (and move into __tests__ folders)
