+++ config/test/test.config.js
@@ -34,6 +34,10 @@ module.exports = {
         "jest-extended/all",
         "<rootDir>/config/test/test-setup.js",
         "<rootDir>/config/test/custom-matchers.js",
+
+        // math-input uses these matchers. We can remove this once we
+        // transition these tests to RTL
+        "<rootDir>/node_modules/jest-enzyme/lib/index.js",
     ],
     moduleNameMapper: {
         "^@khanacademy/perseus(.*)$":
