+++ packages/math-input/package.json
@@ -1,75 +1,41 @@
 {
-  "name": "@khanacademy/math-input",
   "publishConfig": {
     "access": "public"
   },
+  "name": "@khanacademy/math-input",
   "version": "0.4.0",
-  "description": "",
-  "main": "build/math-input.js",
-  "files": [
-    "build/"
-  ],
+  "description": "Khan Academy's new expression editor for the mobile web.",
+  "author": "Khan Academy",
+  "license": "MIT",
+  "module": "dist/es/index.js",
+  "main": "dist/index.js",
+  "source": "src/index.js",
   "scripts": {
-    "test": "jest",
-    "watch_test": "jest --watch",
     "build": "webpack --config webpack.config.externals.js",
     "watch": "webpack --config webpack.config.externals.js --watch",
     "start": "webpack-dev-server --open",
-    "format": "prettier --write \"src/**\" \"test/**\" \"lib/**\"",
-    "prepare": "install-peers",
-    "prepublishOnly": "yarn build",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
-  "author": "",
-  "license": "MIT",
   "dependencies": {
     "performance-now": "^0.2.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/plugin-transform-template-literals": "^7.8.3",
-    "@babel/polyfill": "^7.0.0",
-    "@babel/preset-env": "^7.4.3",
-    "@babel/preset-flow": "^7.8.3",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.0.0",
-    "@storybook/addon-actions": "^5.3.14",
-    "@storybook/addon-backgrounds": "^5.3.14",
-    "@storybook/addon-knobs": "^5.3.14",
-    "@storybook/addon-links": "^5.3.14",
-    "@storybook/addon-viewport": "^5.3.17",
-    "@storybook/addons": "^5.3.14",
-    "@storybook/react": "^5.3.14",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-loader": "7.1.1",
-    "core-js": "^3.0.1",
-    "css-loader": "^3.4.2",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "exports-loader": "^0.6.3",
-    "file-loader": "^4.0.0",
-    "flow-bin": "^0.166.0",
-    "flow-enums-runtime": "^0.0.6",
-    "html-webpack-plugin": "^3.2.0",
-    "install-peers-cli": "^2.2.0",
-    "jest": "24",
-    "jest-enzyme": "^7.1.2",
-    "jsdom": "^8.4.0",
-    "less": "^2.7.2",
-    "less-loader": "^5.0.0",
-    "less-plugin-autoprefix": "^1.5.1",
-    "less-plugin-clean-css": "^1.5.1",
-    "mini-css-extract-plugin": "^0.9.0",
-    "prettier": "^1.19.1",
+    "@khanacademy/wonder-blocks-clickable": "^2.2.0",
+    "@khanacademy/wonder-blocks-color": "^1.1.19",
+    "@khanacademy/wonder-blocks-core": "^3.2.0",
+    "aphrodite": "^1.1.0",
+    "jquery": "^2.1.1",
+    "katex": "^0.11.1",
+    "mathquill": "git+https://git@github.com/Khan/mathquill.git#a9ae54e057c5c1acc8244a5627acbff29901d992",
+    "prop-types": "15.6.1",
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack": "^4.32.2",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.4.1"
+    "react-redux": "^7.2.6",
+    "react-router": "^5.2.1",
+    "react-router-dom": "^5.3.0",
+    "react-transition-group": "^4.4.1",
+    "redux": "^4.0.0"
   },
   "peerDependencies": {
     "@khanacademy/wonder-blocks-clickable": "^2.2.0",
