+++ packages/math-input/src/components/tabbar/__tests__/tabbar_test.js
@@ -1,8 +1,7 @@
-import React from "react";
-import assert from "assert";
 import {mount} from "enzyme";
+import React from "react";
 
-import Tabbar from "../src/components/tabbar/tabbar";
+import Tabbar from "../tabbar.js";
 
 describe("<Tabbar />", () => {
     it("defaults to selecting the first item", () => {
